feat: Convert company items page to server component + Server Actions
- Convert /dashboard/company-items from client component to server component
- Add CompanyItemsClient for interactive functionality
- Remove obsolete /api/evaluation-items/all endpoint
- Update API audit: 9 of 21 endpoints eliminated (43% reduction achieved)

Technical Implementation:
- Server component with direct database queries using Prisma
- Client component handles form interactions and state management
- Progressive enhancement with server-side data fetching
- Maintains all existing functionality while removing API dependency
- Proper data serialization for server-to-client communication

API Simplification Progress:
- Phase 1 (Server Components): 3 endpoints eliminated
- Phase 2 (Server Actions): 7 endpoints eliminated
- Total progress: 43% of original API endpoints removed
- Remaining: 6 convertible GET endpoints (complex interactive forms)
