--- src/app/api/evaluation-items/all/route.ts
@@ -1,91 +1,0 @@
-import { NextResponse } from 'next/server'
-import { auth } from '@/auth'
-import { prisma } from '@/lib/prisma-client'
-
-// GET /api/evaluation-items/all - Get all evaluation items with assignment info
-export async function GET(request: Request) {
-  try {
-    const session = await auth()
-    
-    if (!session?.user?.id) {
-      return NextResponse.json({ 
-        success: false, 
-        error: 'Unauthorized' 
-      }, { status: 401 })
-    }
-
-    const companyId = session.user.companyId
-    const { searchParams } = new URL(request.url)
-    const includeInactive = searchParams.get('includeInactive') === 'true'
-
-    // Get evaluation items for the company
-    const whereClause: { companyId: string; active?: boolean } = { companyId }
-    if (!includeInactive) {
-      whereClause.active = true
-    }
-
-    const items = await prisma.evaluationItem.findMany({
-      where: whereClause,
-      include: {
-        creator: {
-          select: {
-            name: true,
-            role: true
-          }
-        },
-        deadlineSetByUser: {
-          select: {
-            name: true,
-            role: true
-          }
-        },
-        individualAssignments: {
-          include: {
-            employee: {
-              select: {
-                name: true,
-                email: true,
-                username: true
-              }
-            }
-          }
-        }
-      },
-      orderBy: [
-        { level: 'asc' }, // company, department, manager
-        { sortOrder: 'asc' }
-      ]
-    })
-
-    // Transform to include assignment information and deadlines
-    const formattedItems = items.map((item) => ({
-      id: item.id,
-      title: item.title,
-      description: item.description,
-      type: item.type,
-      level: item.level,
-      active: item.active, // ✅ Include active status
-      createdBy: item.creator.name,
-      creatorRole: item.creator.role,
-      assignedEmployees: item.individualAssignments.map(assignment => assignment.employee.name),
-      assignedEmployeeIds: item.individualAssignments.map(assignment => assignment.employeeId),
-      evaluationDeadline: item.evaluationDeadline?.toISOString() || null,
-      deadlineSetBy: item.deadlineSetByUser?.name || null,
-      deadlineSetByRole: item.deadlineSetByUser?.role || null,
-      assignedTo: item.assignedTo,
-      createdAt: item.createdAt.toISOString() // ✅ Also include creation date for sorting
-    }))
-
-    return NextResponse.json({ 
-      success: true,
-      items: formattedItems 
-    })
-  } catch (error) {
-    console.error('Error fetching all evaluation items:', error)
-    return NextResponse.json({ 
-      success: false,
-      error: 'Failed to fetch evaluation items',
-      details: error instanceof Error ? error.message : 'Unknown error'
-    }, { status: 500 })
-  }
-}
